requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2; python_version >= '3.7'
cachetools==5.3.1; python_version >= '3.7'
certifi==2023.7.22; python_version >= '3.6'
charset-normalizer==3.2.0; python_full_version >= '3.7.0'
click==8.1.6; python_version >= '3.7'
dj-database-url==0.5.0
django==3.2.20
flask==2.0.3
google-api-core[grpc]==2.11.1; python_version >= '3.7'
google-auth==2.22.0; python_version >= '3.6'
google-cloud-tasks==2.13.1
googleapis-common-protos==1.60.0; python_version >= '3.7'
grpc-google-iam-v1==0.12.6; python_version >= '3.7'
grpcio==1.57.0
grpcio-status==1.57.0
gunicorn==20.1.0
idna==3.4; python_version >= '3.5'
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
markupsafe==2.1.3; python_version >= '3.7'
proto-plus==1.22.3; python_version >= '3.6'
protobuf==4.24.0; python_version >= '3.7'
pyasn1==0.5.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pyasn1-modules==0.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pytz==2023.3
requests==2.31.0; python_version >= '3.7'
rsa==4.9; python_version >= '3.6' and python_version < '4'
setuptools==68.0.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.4; python_version >= '3.5'
typing-extensions==4.7.1; python_version < '3.11'
urllib3==1.26.16; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
werkzeug==2.3.6; python_version >= '3.8'
